src/app/requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-pinecone
Unstructured
pypdf
pdfminer.six
poppler-utils
pillow_heif
pdf2image
unstructured_inference
pytesseract
pinecone-client
pinecone
tiktoken
python-dotenv
datasets
pandas
ragas
tqdm
rank_bm25
giskard
jupyterlab
tenacity
streamlit
pydantic
neo4j
neo4j-driver
textblob
colorama

# Key: sk-proj-vg4XfnAxDqOrdiStgxr0T3BlbkFJcTOGxrEt1taQVpFseL0E
